Fix: silence -Wsign-compare in dkim-canon.c
dkim_canon_flushblanks: change loop counter c from int to u_int to
match canon->canon_blanks.

dkim_add_canon: cast the int hashtype parameter to u_int at the
comparison with struct field canon_hashtype (u_int).  hashtype is
asserted to be a DKIM_HASHTYPE_* enum value upstream, so the cast
is lossless.

+++ libopendkim/dkim-canon.c
@@ -459,7 +459,7 @@ dkim_canon_header(DKIM *dkim, DKIM_CANON *canon, struct dkim_header *hdr,
 static void
 dkim_canon_flushblanks(DKIM_CANON *canon)
 {
-	int c;
+	u_int c;
 
 	assert(canon != NULL);
 
@@ -725,7 +725,7 @@ dkim_add_canon(DKIM *dkim, _Bool hdr, dkim_canon_t canon, int hashtype,
 		for (cur = dkim->dkim_canonhead; cur != NULL; cur = cur->canon_next)
 		{
 			if (cur->canon_hdr ||
-			    cur->canon_hashtype != hashtype ||
+			    cur->canon_hashtype != (u_int) hashtype ||
 			    cur->canon_canon != canon)
 				continue;
